Avances hechos hasta acá. No funca.
Me está volviendo loco esto de las jugada válida.
