+++ private/datos.md
@@ -1,16 +1,11 @@
-""" Este módulo permite jugar al Whist de forma cooperativa local. """
-
 # Qué tipos de datos hay?
 
-""" Datos mutables
- De menor a mayor
 
  - Lista de jugadores por mano
  - Puntos de cada jugador
  - Predicciones de la mano
  - Cartas que tiene cada uno en cada baza
- - Triunfo  """
+ - Triunfo
 
 
-
-# Datos inmutables
+## Datos inmutables
