+++ src/whist/__init__.py
@@ -1,0 +1,5 @@
+""" Este módulo contiene funciones para jugar al Whist de forma cooperativa local. """
+
+import whist.impresion
+import whist.entrada
+import whist.logica
