+++ src/whist/entrada.py
@@ -1,7 +1,8 @@
 """ Este modulo incluye todas las funciones para obtener la entrada del usuario. """
 
 import random
-
+from .impresion import imprimir_error_jugada
+from .logica import corroborar_jugada
 
 def ingresar_jugadores():
     """ Se solicita al usuario inicial que ingrese los nombres de los jugadores separados por
@@ -51,44 +52,11 @@ def ingresar_jugada(jugador: str, cartas_jugador: "list[tuple]",
                 input(f"Debe seleccionar una carta dentro del rango 1-{largo}.")
             elif not palo_baza:
                 condicion = False
-            elif corroborar_jugada(cartas_jugador, jugada, palo_baza[1], triunfo[1]):
+            jugada = corroborar_jugada(cartas_jugador, jugada, palo_baza[1], triunfo[1])
+            if jugada.validez:
                 condicion = False
+            else:
+                imprimir_error_jugada(jugada.tipo, palo_baza[1], jugada.cartas)
         except ValueError:
             input("Debe ingresar un número.")
     return cartas_jugador[jugada-1]
-        
-
-def corroborar_jugada(cartas_jugador: "list[tuple]", jugada: int,
-                      palo_baza: str, palo_triunfo: str) -> bool:
-
-    carta_seleccionada = cartas_jugador[jugada-1]
-    palo_carta = carta_seleccionada[1]
-    tiene_palo = False
-    tiene_triunfo = False
-    palos_baza_disponibles: list[int] = list()
-    palos_triunfo_disponibles: list[int] = list()
-    if palo_carta == palo_baza:
-        return True
-    for numero, carta in enumerate(cartas_jugador):
-        if carta[1] == palo_baza:
-            tiene_palo = True
-            palos_baza_disponibles.append(numero+1)
-        if carta[1] == palo_triunfo:
-            tiene_triunfo = True
-            palos_triunfo_disponibles.append(numero+1)
-    if not tiene_palo and palo_carta == palo_triunfo:
-        return True
-    if not tiene_palo and not tiene_triunfo:
-        return True
-    if tiene_palo:
-        print(f"Seleccionó una carta con un palo distinto al de la baza teniendo "
-              f"al menos una carta con dicho palo ('{palo_baza}').")
-        print("Debe seleccionar alguna de sus cartas con el palo de la baza.\n"
-              f"Las que cumplen con la condición son las numero {palos_baza_disponibles}.")
-    else:
-        print(f"Seleccionó una carta con un palo distinto al de triunfo teniendo "
-              f"al menos una carta con dicho palo ('{palo_triunfo}').")
-        print("Debe seleccionar alguna de sus cartas con el palo de triunfo, ya que no "
-                "tiene ninguna con el palo de la baza.\n"
-                f"Las que cumplen con la condición son las numero {palos_triunfo_disponibles}.")
-    return False
